+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # react-parm CHANGELOG
 
+## 2.0.0
+
+#### BREAKING CHANGES
+
+* Component functions used with `createComponent` now uses `createRender`, which accepts props as the first argument (not the full instance)
+
+#### NEW FEATURES
+
+* Add [`createRender`](README.md#createrender) method
+
 ## 1.1.1
 
 * Use ES5 version of component class extension (smaller footprint)